Adding typecast double to int while calling std::sqrt for compiling Visual Studio 2010

+++ modules/text/src/ocr_beamsearch_decoder.cpp
@@ -517,7 +517,7 @@ OCRBeamSearchClassifierCNN::OCRBeamSearchClassifierCNN (const string& filename)
 
     nr_feature = weights.rows;
     nr_class   = weights.cols;
-    patch_size  = (int)sqrt(kernels.cols);
+    patch_size  = (int)sqrt((double)kernels.cols);
     window_size = 4*patch_size;
     step_size   = 4;
     quad_size   = 12;
